Move entities and repositories directly under persistence package
