+++ src/main/java/lt/vu/application/address/factory/AddressFactory.java
@@ -1,6 +1,6 @@
 package lt.vu.application.address.factory;
 
-import lt.vu.persistence.orm.entities.Address;
+import lt.vu.persistence.entities.Address;
 import lt.vu.web.api.v1.dto.address.PostAddressDTO;
 
 import javax.enterprise.context.RequestScoped;
