+++ src/main/java/lt/vu/application/userInfo/factory/UserInfoFactory.java
@@ -1,7 +1,7 @@
 package lt.vu.application.userInfo.factory;
 
 import lt.vu.application.address.factory.AddressFactory;
-import lt.vu.persistence.orm.entities.UserInfo;
+import lt.vu.persistence.entities.UserInfo;
 import lt.vu.application.user.service.UserInfoDTO;
 
 import javax.enterprise.context.RequestScoped;
