+++ src/main/java/lt/vu/persistence/entities/AttributeType.java
@@ -1,4 +1,4 @@
-package lt.vu.persistence.orm.entities;
+package lt.vu.persistence.entities;
 
 /**
  * Allows to identify different attributes by purpose inside the backend.
