+++ src/main/java/lt/vu/persistence/entities/PackageSize.java
@@ -1,4 +1,4 @@
-package lt.vu.persistence.orm.entities;
+package lt.vu.persistence.entities;
 
 import lombok.Getter;
 import lombok.Setter;
