+++ src/main/java/lt/vu/persistence/repository/PackageOptionRepository.java
@@ -1,8 +1,8 @@
-package lt.vu.persistence.orm.repository;
+package lt.vu.persistence.repository;
 
 import lt.vu.application.exception.NotFoundException;
 import lt.vu.application.packageOption.exception.PackageOptionNotFoundException;
-import lt.vu.persistence.orm.entities.PackageOption;
+import lt.vu.persistence.entities.PackageOption;
 
 import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
