+++ src/main/java/lt/vu/persistence/repository/PackageSizeRepository.java
@@ -1,6 +1,6 @@
-package lt.vu.persistence.orm.repository;
+package lt.vu.persistence.repository;
 
-import lt.vu.persistence.orm.entities.PackageSize;
+import lt.vu.persistence.entities.PackageSize;
 
 import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
