+++ src/main/java/lt/vu/web/api/v1/admin/controller/order/ListOrderController.java
@@ -6,10 +6,10 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import lt.vu.application.exception.NotFoundException;
 import lt.vu.infrastructure.security.Authorized;
-import lt.vu.persistence.orm.entities.Order;
-import lt.vu.persistence.orm.entities.UserRole;
-import lt.vu.persistence.orm.repository.OrderRepository;
-import lt.vu.persistence.orm.repository.UserRepository;
+import lt.vu.persistence.entities.Order;
+import lt.vu.persistence.entities.UserRole;
+import lt.vu.persistence.repository.OrderRepository;
+import lt.vu.persistence.repository.UserRepository;
 import lt.vu.web.api.v1.dto.order.ListOrderDTO;
 import lt.vu.web.api.v1.dto.order.GetOrderDTO;
 import lt.vu.web.api.v1.exception.ExceptionDTO;
