+++ src/main/java/lt/vu/web/api/v1/admin/controller/order/PutOrderStatusController.java
@@ -8,10 +8,10 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import lt.vu.application.exception.BadRequestException;
 import lt.vu.application.exception.NotFoundException;
 import lt.vu.infrastructure.security.Authorized;
-import lt.vu.persistence.orm.entities.Order;
-import lt.vu.persistence.orm.entities.OrderStatus;
-import lt.vu.persistence.orm.entities.UserRole;
-import lt.vu.persistence.orm.repository.OrderRepository;
+import lt.vu.persistence.entities.Order;
+import lt.vu.persistence.entities.OrderStatus;
+import lt.vu.persistence.entities.UserRole;
+import lt.vu.persistence.repository.OrderRepository;
 import lt.vu.web.api.v1.admin.dto.order.PutOrderStatusDTO;
 import lt.vu.web.api.v1.dto.order.GetOrderDTO;
 import lt.vu.web.api.v1.dto.security.GetTokenDTO;
