+++ src/main/java/lt/vu/web/api/v1/admin/controller/user/ListUserController.java
@@ -5,9 +5,9 @@ import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import lt.vu.infrastructure.security.Authorized;
-import lt.vu.persistence.orm.entities.User;
-import lt.vu.persistence.orm.entities.UserRole;
-import lt.vu.persistence.orm.repository.UserRepository;
+import lt.vu.persistence.entities.User;
+import lt.vu.persistence.entities.UserRole;
+import lt.vu.persistence.repository.UserRepository;
 import lt.vu.web.api.v1.admin.dto.user.GetEnhancedUserDTOFactory;
 import lt.vu.web.api.v1.admin.dto.user.ListEnhancedUserDTO;
 import lt.vu.web.api.v1.exception.ExceptionDTO;
