+++ src/main/java/lt/vu/web/api/v1/controller/attribute/ListAttributeController.java
@@ -4,8 +4,8 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import lt.vu.persistence.orm.entities.Attribute;
-import lt.vu.persistence.orm.repository.AttributeRepository;
+import lt.vu.persistence.entities.Attribute;
+import lt.vu.persistence.repository.AttributeRepository;
 import lt.vu.web.api.v1.dto.attribute.GetAttributeDTO;
 import lt.vu.web.api.v1.dto.attribute.ListAttributeDTO;
