+++ src/main/java/lt/vu/web/api/v1/controller/order/PostOrderController.java
@@ -6,8 +6,8 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.parameters.RequestBody;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import lt.vu.application.exception.NotFoundException;
-import lt.vu.persistence.orm.entities.Order;
-import lt.vu.persistence.orm.repository.OrderRepository;
+import lt.vu.persistence.entities.Order;
+import lt.vu.persistence.repository.OrderRepository;
 import lt.vu.web.api.v1.controller.security.CurrentUserAwareController;
 import lt.vu.web.api.v1.exception.ExceptionDTO;
 import lt.vu.application.order.factory.OrderFactory;
