+++ src/main/java/lt/vu/web/api/v1/controller/packageOption/ListPackageOptionByPackageTypeController.java
@@ -4,8 +4,8 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import lt.vu.persistence.orm.entities.PackageOption;
-import lt.vu.persistence.orm.repository.PackageOptionRepository;
+import lt.vu.persistence.entities.PackageOption;
+import lt.vu.persistence.repository.PackageOptionRepository;
 import lt.vu.web.api.v1.dto.packageOption.ListPackageOptionDTO;
 import lt.vu.web.api.v1.dto.packageOption.GetPackageOptionDTO;
 import lt.vu.web.api.v1.exception.ExceptionDTO;
