+++ src/main/java/lt/vu/web/api/v1/controller/packageSize/ListPackageSizeController.java
@@ -4,8 +4,8 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import lt.vu.persistence.orm.entities.PackageSize;
-import lt.vu.persistence.orm.repository.PackageSizeRepository;
+import lt.vu.persistence.entities.PackageSize;
+import lt.vu.persistence.repository.PackageSizeRepository;
 import lt.vu.web.api.v1.dto.packageSize.ListPackageSizeDTO;
 import lt.vu.web.api.v1.dto.packageSize.GetPackageSizeDTO;
 import lt.vu.web.api.v1.exception.ExceptionDTO;
