+++ src/main/java/lt/vu/web/api/v1/controller/packageType/ListPackageTypeController.java
@@ -4,8 +4,8 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import lt.vu.persistence.orm.entities.PackageType;
-import lt.vu.persistence.orm.repository.PackageTypeRepository;
+import lt.vu.persistence.entities.PackageType;
+import lt.vu.persistence.repository.PackageTypeRepository;
 import lt.vu.web.api.v1.dto.packageType.ListPackageTypeDTO;
 import lt.vu.web.api.v1.dto.packageType.GetPackageTypeDTO;
 import lt.vu.web.api.v1.exception.ExceptionDTO;
