+++ src/main/java/lt/vu/web/api/v1/controller/security/CurrentUserAwareController.java
@@ -1,7 +1,7 @@
 package lt.vu.web.api.v1.controller.security;
 
 import lt.vu.infrastructure.security.LoggedInUser;
-import lt.vu.persistence.orm.entities.User;
+import lt.vu.persistence.entities.User;
 
 import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
