+++ src/main/java/lt/vu/web/api/v1/dto/packageSize/GetPackageSizeDTO.java
@@ -2,7 +2,7 @@ package lt.vu.web.api.v1.dto.packageSize;
 
 import lombok.Getter;
 import lombok.Setter;
-import lt.vu.persistence.orm.entities.PackageSize;
+import lt.vu.persistence.entities.PackageSize;
 
 import java.util.List;
 import java.util.stream.Collectors;
